Rearranged code, created lib.rs
