--- src/entrypoint/unix/mod.rs
@@ -1,3 +1,0 @@
-mod unix;
-mod promptbuffer;
-pub use self::unix::main;
